Install only required packages if option is set
The current solution installs all packages found by pip freeze to the
distribution folder and deploys this to AWS Lambda.
This includes some unnecessary packages to run the code, e.g. testing modules
which increase the distribution size.

The option to install only required packages helps reducing the build size.
For now, the file must be named `requirements.txt` and be in the same
directory of the project which shall be uploaded.